% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{get_twitter_token}
\alias{get_twitter_token}
\title{Make connection to Twitter (Deprecated)}
\usage{
get_twitter_token()
}
\value{
A connection to local db
}
\description{
Make connection to Twitter (Deprecated)
}
